sns_trigger_lambda/requirements.txt
SQLAlchemy
python-dotenv
pandas